+++ src/code42cli/config.py
@@ -1,6 +1,7 @@
 from __future__ import print_function
 
 import os
+
 from configparser import ConfigParser
 
 import code42cli.util as util